+++ src/models/options/html-converter-options.ts
@@ -1,6 +1,6 @@
 export class HtmlConverterOptions {
   /**
-   * if set true, adds the table and image widthWithUnit property.
+   * If set true, adds the table and image widthWithUnit property.
    * Sample output will be ' width: 100,  widthWithUnit: { value: 20, unit: "Percentage" } '.
    * The default value is false, which does not have widthWithUnit property.
    * The width property is always in em.
@@ -11,9 +11,14 @@ export class HtmlConverterOptions {
    * If user needs to override, then use this option like 'baseFontSize : 20'.
    */
   baseFontSize?: number = 16;
+  /**
+   * Relative hyperlink paths will be extended to absolute with the base url
+   */
+  hyperlinkBaseUrl?: string = '';
 }
 
 export const DefaultOptions: HtmlConverterOptions = {
   handleWidthWithUnits: false,
   baseFontSize: 16,
+  hyperlinkBaseUrl: '',
 };